add tests for validation registry
